requirements/base.txt
Django==1.6.3
bpython==0.12
django-braces==1.2.2
django-model-utils==1.5.0
logutils==0.3.3
South==0.8.2

django-geoposition==0.1.5
django-zeropush==0.2.3
paypalrestsdk==1.1.0
psycopg2==2.5.2
pymongo==2.7
pytz==2014.1.1
requests==2.2.1